requirements.txt
fastapi
uvicorn[standard]
sqlmodel
Jinja2
python-multipart
alembic
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
